refactor(platform): split posix and windows backends
