+++ Makefile.in
@@ -135,7 +135,8 @@ PROGRAMS = $(bin_PROGRAMS)
 am_spine_OBJECTS = sql.$(OBJEXT) spine.$(OBJEXT) util.$(OBJEXT) \
 	snmp.$(OBJEXT) locks.$(OBJEXT) poller.$(OBJEXT) \
 	nft_popen.$(OBJEXT) php.$(OBJEXT) ping.$(OBJEXT) \
-	keywords.$(OBJEXT) error.$(OBJEXT) platform.$(OBJEXT)
+	keywords.$(OBJEXT) error.$(OBJEXT) platform_common.$(OBJEXT) \
+	platform_posix.$(OBJEXT) platform_win.$(OBJEXT)
 spine_OBJECTS = $(am_spine_OBJECTS)
 spine_LDADD = $(LDADD)
 AM_V_lt = $(am__v_lt_@AM_V@)
@@ -159,9 +160,11 @@ depcomp = $(SHELL) $(top_srcdir)/config/depcomp
 am__maybe_remake_depfiles = depfiles
 am__depfiles_remade = ./$(DEPDIR)/error.Po ./$(DEPDIR)/keywords.Po \
 	./$(DEPDIR)/locks.Po ./$(DEPDIR)/nft_popen.Po \
-	./$(DEPDIR)/php.Po ./$(DEPDIR)/ping.Po ./$(DEPDIR)/platform.Po \
-	./$(DEPDIR)/poller.Po ./$(DEPDIR)/snmp.Po ./$(DEPDIR)/spine.Po \
-	./$(DEPDIR)/sql.Po ./$(DEPDIR)/util.Po
+	./$(DEPDIR)/php.Po ./$(DEPDIR)/ping.Po \
+	./$(DEPDIR)/platform_common.Po ./$(DEPDIR)/platform_posix.Po \
+	./$(DEPDIR)/platform_win.Po ./$(DEPDIR)/poller.Po \
+	./$(DEPDIR)/snmp.Po ./$(DEPDIR)/spine.Po ./$(DEPDIR)/sql.Po \
+	./$(DEPDIR)/util.Po
 am__mv = mv -f
 COMPILE = $(CC) $(DEFS) $(DEFAULT_INCLUDES) $(INCLUDES) $(AM_CPPFLAGS) \
 	$(CPPFLAGS) $(AM_CFLAGS) $(CFLAGS)
@@ -391,11 +394,11 @@ top_builddir = @top_builddir@
 top_srcdir = @top_srcdir@
 AUTOMAKE_OPTIONS = foreign
 ACLOCAL_AMFLAGS = -I m4
-spine_SOURCES = sql.c spine.c util.c snmp.c locks.c poller.c nft_popen.c php.c ping.c keywords.c error.c platform.c
+spine_SOURCES = sql.c spine.c util.c snmp.c locks.c poller.c nft_popen.c php.c ping.c keywords.c error.c platform_common.c platform_posix.c platform_win.c
 configdir = $(sysconfdir)
 config_DATA = spine.conf.dist
 man_MANS = spine.1
-EXTRA_DIST = spine.1 uthash.h spine_sem.h platform.h tests/unit/Makefile tests/unit/test_platform_smoke.c
+EXTRA_DIST = spine.1 uthash.h spine_sem.h platform.h platform_common.c platform_posix.c platform_win.c tests/unit/Makefile tests/unit/test_platform_smoke.c
 all: all-am
 
 .SUFFIXES:
@@ -509,7 +512,9 @@ distclean-compile:
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/nft_popen.Po@am__quote@ # am--include-marker
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/php.Po@am__quote@ # am--include-marker
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/ping.Po@am__quote@ # am--include-marker
-@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/platform.Po@am__quote@ # am--include-marker
+@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/platform_common.Po@am__quote@ # am--include-marker
+@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/platform_posix.Po@am__quote@ # am--include-marker
+@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/platform_win.Po@am__quote@ # am--include-marker
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/poller.Po@am__quote@ # am--include-marker
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/snmp.Po@am__quote@ # am--include-marker
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/spine.Po@am__quote@ # am--include-marker
@@ -902,7 +907,9 @@ distclean: distclean-am
 	-rm -f ./$(DEPDIR)/nft_popen.Po
 	-rm -f ./$(DEPDIR)/php.Po
 	-rm -f ./$(DEPDIR)/ping.Po
-	-rm -f ./$(DEPDIR)/platform.Po
+	-rm -f ./$(DEPDIR)/platform_common.Po
+	-rm -f ./$(DEPDIR)/platform_posix.Po
+	-rm -f ./$(DEPDIR)/platform_win.Po
 	-rm -f ./$(DEPDIR)/poller.Po
 	-rm -f ./$(DEPDIR)/snmp.Po
 	-rm -f ./$(DEPDIR)/spine.Po
@@ -961,7 +968,9 @@ maintainer-clean: maintainer-clean-am
 	-rm -f ./$(DEPDIR)/nft_popen.Po
 	-rm -f ./$(DEPDIR)/php.Po
 	-rm -f ./$(DEPDIR)/ping.Po
-	-rm -f ./$(DEPDIR)/platform.Po
+	-rm -f ./$(DEPDIR)/platform_common.Po
+	-rm -f ./$(DEPDIR)/platform_posix.Po
+	-rm -f ./$(DEPDIR)/platform_win.Po
 	-rm -f ./$(DEPDIR)/poller.Po
 	-rm -f ./$(DEPDIR)/snmp.Po
 	-rm -f ./$(DEPDIR)/spine.Po
